Split multi string form field and string form field
